This is Daily News Web App developed from HTML5,CSS3,JAVASCRIPT,AJAX,BOOTSTRAP,JQUERY. In this web app the user can browse through the news headlines of different categories like Business, Technology, General etc. The news feed will be refreshed every time so that the user can see the latest and top news and be up to date with the world.
This is Daily News Web App developed from HTML5,CSS3,JAVASCRIPT,AJAX,BOOTSTRAP,JSON and JQUERY. The fronend of this web app is managed with the help of these technologies where the user sees the latest news that are returned from the News API. In the backend of this web app an AJAX call is made with the help of JQUERY and then the News API return the specific parameters in the form of JSON (Javascript Object Notation) which is then parsed and the corresponding results are shown to the user in the front end. Every time the user refreshes the page the corresponding API call is made.